feature : student add and view details

+++ backend/routes/Student.js
@@ -5,15 +5,14 @@ const Batch = require("../models/Batch.js");
 const userAuth  =require("../middleware/userAuth.js");
 
 
-router.post("/add-new-student/:batchId",userAuth, async (req, res) => {
-    const batchId = req.params.batchId;
-    const { contact_info } = req.body;
+router.post("/add-new-student",userAuth, async (req, res) => {
+    const { contact_info, batchId } = req.body;
     const adminId = req.user._id;
 
     try {
         const existingStudent = await Student.findOne({
             adminId:adminId,
-            bathId:batchId,
+            batchId:batchId,
             $or: [
                 { "contact_info.emailIds.student": contact_info.emailIds.student },
                 { "contact_info.phoneNumbers.student": contact_info.phoneNumbers.student }
@@ -42,32 +41,37 @@ router.post("/add-new-student/:batchId",userAuth, async (req, res) => {
         return res.status(500).json({ message: "Failed to add student", error: error.message });
     }
 });
-router.get("/get-all-students-of-institute",userAuth, async (req, res) => {
-    try{
-        const adminId = req.user._id;
-        const response = await Student.find({adminId: adminId}).populate('admin');
-        console.log(response);
-        return res.status(200).json(response);
 
-    }catch(error){
-        console.error("Error adding student:", error);
+router.get("/get-all-students-of-institute", userAuth, async (req, res) => {
+    try {
+        const adminId = req.user._id;
+        const students = await Student.find({ adminId }).populate('batchId');
+        return res.status(200).json(students);
+    } catch (error) {
+        console.error("Error fetching students:", error);
         return res.status(500).json({ message: "Failed to fetch students", error: error.message });
     }
-})
+});
+
+
 router.delete("/delete-student/:id",userAuth, async (req, res) => {
     try {
         const { id } = req.params;
-        const response = await Student.deleteOne({ _id: id });
+        const adminId = req.user._id;
+        const response = await Student.deleteOne({ adminId ,_id: id });
         console.log(response);
         return res.status(200).json(response.data);
     }catch(error){
         console.error("Error deleting student:", error.message);
     }
 })
-router.put("/update-student/:id", userAuth,async (req, res) => {
+
+router.patch("/update-student/:id", userAuth,async (req, res) => {
     const { id } = req.params;
+    const adminId = req.user._id;
+
     try {
-        const updated = await Student.findByIdAndUpdate(id, req.body, { new: true });
+        const updated = await Student.findByIdAndUpdate({id, adminId}, req.body, { new: true });
         if (!updated) return res.status(404).json({ message: `${name} not found` });
 
         res.status(200).json(updated);
